コントローラー編集
コントローラーのdestroyに出品者とログイン者が同じかどうかの分岐を加えた

+++ app/controllers/items_controller.rb
@@ -25,6 +25,7 @@ class ItemsController < ApplicationController
   end
 
   def destroy
+    return unless current_user == @item.user
     return unless @item.destroy
 
     redirect_to root_path, notice: '削除されました。'
